Made a method to convert the instructions into machine code.
There are a few different types of instructions, so the code looks at the part of the opcode that differentiates the different instructions and applies the needed computation and conversion to make the machine code.
Requires there to be no pseudo operations in the instruction list, as it is a direct conversion from instruction to machine code.